requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
mypy==1.15.0
types-requests==2.31.0.10